submit_service/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
httpx>=0.27.0
pydantic>=2.7.0
PyYAML>=6.0
jinja2>=3.1.0
python-multipart>=0.0.9
